Core/add more engineering tests (#14)
* half the programs work

* tried to take out prereqs not in program

* Added logic to handle pre-req edge cases

* Added logic to handle pre-req edge cases

* added civil engineering, and removed the "taken" patch since the removal of prereqs outside  of the program fixes that issue

* Removed old server folder structure

* added static version of main method

+++ Common/src/main/java/skynet/scheduler/common/Course.java
@@ -28,7 +28,7 @@ public class Course implements ICourse
     		String[] prerequisites,
     		String courseSubject,
     		String courseCatalog,
-    		int creditUnits,
+    		double creditUnits,
     		String academicCareer) 
     {
         this.courseName = courseName;
